+++ test/rest/2023_07/shop_test.rb
@@ -91,4 +91,22 @@ class Shop202307Test < Test::Unit::TestCase
     end
   end
 
+  sig do
+    void
+  end
+  def test_current
+    stub_request(:get, "https://test-shop.myshopify.io/admin/api/2023-07/shop.json")
+      .with(
+        headers: {"X-Shopify-Access-Token"=>"this_is_a_test_token", "Accept"=>"application/json"},
+        body: {}
+      )
+      .to_return(status: 200, body: JSON.generate({"shop" => {"province" => "California", "country" => "US", "address1" => "1 Infinite Loop", "city" => "Cupertino", "address2" => "Suite 100"}}), headers: {})
+
+    shop = ShopifyAPI::Shop.current
+
+    assert_requested(:get, "https://test-shop.myshopify.io/admin/api/2023-07/shop.json")
+
+    assert_equal(shop.province, "California")
+  end
+
 end